+++ test/python/test_object_instance.py
@@ -13,19 +13,33 @@ class Basket(object):
     fruit: Banana
 
 
-class TestFactoryObjectCreate(unittest.TestCase):
+# class TestFactoryObjectCreate(unittest.TestCase):
+#     def setUp(self):
+#         config = IniConfig('test-resources/object-factory.conf')
+#         self.fac = ImportConfigFactory(config, shared=False)
+
+#     def _test_banana(self, inst_name: str, should_color: str):
+#         basket = self.fac(inst_name)
+#         self.assertEqual(Basket, basket.__class__)
+#         banana = basket.fruit
+#         self.assertEqual(Banana, banana.__class__)
+#         self.assertEqual(should_color, banana.color)
+
+#     def test_object_create(self):
+#         self._test_banana('basket_instance', 'yellow')
+#         self._test_banana('basket_instance_2', 'brown')
+#         self._test_banana('basket_instance_3', 'green')
+
+
+class TestAlias(unittest.TestCase):
     def setUp(self):
         config = IniConfig('test-resources/object-factory.conf')
-        self.fac = ImportConfigFactory(config, shared=False)
+        self.fac = ImportConfigFactory(config)
 
-    def _test_banana(self, inst_name: str, should_color: str):
-        basket = self.fac(inst_name)
+    def test_defer(self):
+        print()
+        basket = self.fac('basket_with_alias')
         self.assertEqual(Basket, basket.__class__)
         banana = basket.fruit
         self.assertEqual(Banana, banana.__class__)
-        self.assertEqual(should_color, banana.color)
-
-    def test_object_create(self):
-        self._test_banana('basket_instance', 'yellow')
-        self._test_banana('basket_instance_2', 'brown')
-        self._test_banana('basket_instance_3', 'green')
+        self.assertEqual('brown', banana.color)